ebpf_plugin: Rename files
This commit rename files to the final version
